Move methods around and create arwen/context/execution.go
