+++ arwen/helpers_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/ElrondNetwork/arwen-wasm-vm/arwen/context"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -19,7 +20,7 @@ func Test_TryCatch_WorksWhenNoError(t *testing.T) {
 		catchCalled = true
 	}
 
-	TryCatch(try, catch, "message")
+	context.TryCatch(try, catch, "message")
 
 	assert.True(t, tryCalled)
 	assert.False(t, catchCalled)
@@ -38,7 +39,7 @@ func Test_TryCatch_CatchesRuntimeError(t *testing.T) {
 		caughtError = err
 	}
 
-	TryCatch(try, catch, "message")
+	context.TryCatch(try, catch, "message")
 
 	assert.NotNil(t, caughtError)
 }
@@ -54,7 +55,7 @@ func Test_TryCatch_CatchesCustomError(t *testing.T) {
 		caughtError = err
 	}
 
-	TryCatch(try, catch, "!thisMessage!")
+	context.TryCatch(try, catch, "!thisMessage!")
 
 	assert.NotNil(t, caughtError)
 	assert.Contains(t, caughtError.Error(), "!thisMessage!")
@@ -73,7 +74,7 @@ func Test_TryCatch_CatchesCustomErrorTyped(t *testing.T) {
 		caughtError = err
 	}
 
-	TryCatch(try, catch, "!thisMessage!")
+	context.TryCatch(try, catch, "!thisMessage!")
 
 	assert.NotNil(t, caughtError)
 	assert.Equal(t, customError, caughtError)
